+++ rustls/src/suites.rs
@@ -1,4 +1,4 @@
-use crate::cipher::{self, Tls12AeadAlgorithm};
+use crate::tls12;
 use crate::msgs::enums::ProtocolVersion;
 use crate::msgs::enums::{CipherSuite, SignatureAlgorithm, SignatureScheme};
 use crate::msgs::handshake::DecomposedSignatureScheme;
@@ -116,7 +116,7 @@ pub struct Tls12CipherSuite {
     /// chacha20poly1305 works this way by design.
     pub explicit_nonce_len: usize,
 
-    pub(crate) aead_alg: &'static dyn Tls12AeadAlgorithm,
+    pub(crate) aead_alg: &'static dyn tls12::Tls12AeadAlgorithm,
 }
 
 impl Tls12CipherSuite {
@@ -235,7 +235,7 @@ pub static TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256: SupportedCipherSuite =
         sign: TLS12_ECDSA_SCHEMES,
         fixed_iv_len: 12,
         explicit_nonce_len: 0,
-        aead_alg: &cipher::ChaCha20Poly1305,
+        aead_alg: &tls12::ChaCha20Poly1305,
         hmac_algorithm: ring::hmac::HMAC_SHA256,
     });
 
@@ -251,7 +251,7 @@ pub static TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256: SupportedCipherSuite =
         sign: TLS12_RSA_SCHEMES,
         fixed_iv_len: 12,
         explicit_nonce_len: 0,
-        aead_alg: &cipher::ChaCha20Poly1305,
+        aead_alg: &tls12::ChaCha20Poly1305,
         hmac_algorithm: ring::hmac::HMAC_SHA256,
     });
 
@@ -267,7 +267,7 @@ pub static TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256: SupportedCipherSuite =
         sign: TLS12_RSA_SCHEMES,
         fixed_iv_len: 4,
         explicit_nonce_len: 8,
-        aead_alg: &cipher::AesGcm,
+        aead_alg: &tls12::AesGcm,
         hmac_algorithm: ring::hmac::HMAC_SHA256,
     });
 
@@ -283,7 +283,7 @@ pub static TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384: SupportedCipherSuite =
         sign: TLS12_RSA_SCHEMES,
         fixed_iv_len: 4,
         explicit_nonce_len: 8,
-        aead_alg: &cipher::AesGcm,
+        aead_alg: &tls12::AesGcm,
         hmac_algorithm: ring::hmac::HMAC_SHA384,
     });
 
@@ -299,7 +299,7 @@ pub static TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256: SupportedCipherSuite =
         sign: TLS12_ECDSA_SCHEMES,
         fixed_iv_len: 4,
         explicit_nonce_len: 8,
-        aead_alg: &cipher::AesGcm,
+        aead_alg: &tls12::AesGcm,
         hmac_algorithm: ring::hmac::HMAC_SHA256,
     });
 
@@ -315,7 +315,7 @@ pub static TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384: SupportedCipherSuite =
         sign: TLS12_ECDSA_SCHEMES,
         fixed_iv_len: 4,
         explicit_nonce_len: 8,
-        aead_alg: &cipher::AesGcm,
+        aead_alg: &tls12::AesGcm,
         hmac_algorithm: ring::hmac::HMAC_SHA384,
     });
